api/pyproject.toml
[project]
name = "chinmey-farms-services"
version = "1.0.0"
description = ""
authors = [
    {name = "subramanyam.mogili"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12,<0.116.0",
    "uvicorn[standard]>=0.23.0,<0.24.0",
    "sqlalchemy[asyncio]>=2.0.40,<3.0.0",
    "asyncpg>=0.29.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[tool.poetry]
packages = []

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
